+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -244,7 +244,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K6D83JS7TBK9AS2GPN622FYB"),
+  _id: z.string().default("01K6DAH8K4MVV16JYD3JV0F9RT"),
   display_name: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$inboundSchema,
@@ -288,7 +288,7 @@ export const UpdateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01K6D83JS7TBK9AS2GPN622FYB"),
+  id: z.string().default("01K6DAH8K4MVV16JYD3JV0F9RT"),
   displayName: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$outboundSchema,