+++ packages/orq-rc/RELEASES.md
@@ -3979,3 +3979,13 @@ Based on:
 - [typescript v3.13.0-rc.50] packages/orq-rc
 ### Releases
 - [NPM v3.13.0-rc.50] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.50 - packages/orq-rc
+
+## 2025-09-30 12:47:43
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.631.5 (2.716.16) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.13.0-rc.51] packages/orq-rc
+### Releases
+- [NPM v3.13.0-rc.51] https://www.npmjs.com/package/@orq-ai/node/v/3.13.0-rc.51 - packages/orq-rc